Add new tests implementation
